+++ src/parser.ts
@@ -104,33 +104,3 @@ export function findResultBlock(
   const foundEndLine = foundStartLine + codeLineCount - 1;
   return new vscode.Range(foundStartLine, 0, foundEndLine, 0);
 }
-
-// Parses out any inline code (code within ` delimiters)
-// This is a generator function that yields the code of each code block
-export function* parseInlineCode(
-  document: vscode.TextDocument
-): Generator<{ code: string; range: vscode.Range }> {
-  // Explanation of regex:
-  // [^`\n] means match all any character that is not ` or \n (negated class)
-  // ([^`\n]+?) capturing group will then match one or more of that class
-  // It will also do so lazily, and not greedily because of the ?
-  // Thus, `([^`\n]+?)` effectively matches an inline code block, with the
-  // capturing group matching the code itself (non ` or \n characters)
-  // Finally, we add some extra validation using look ahead and look behind:
-  // (?<!`+) means negative look behind of at least one `
-  // (?!`+) means negative look ahead of at least one `
-  // This means that if there are multiple consecutive ` before and/or after
-  // the code block, then we reject it.
-  const regex: RegExp = /(?<!`+)`([^`\n]+?)`(?!`+)/g;
-
-  // Loop through all matches and yield them
-  let match;
-  while ((match = regex.exec(document.getText())) !== null) {
-    // match[0] captures the entire code block (we do not need it)
-    const code = match[1]; // First capturing group ([^`\n]+?)
-    const startPos = document.positionAt(match.index);
-    const endPos = document.positionAt(match.index + match[0].length);
-    const range = new vscode.Range(startPos, endPos); // Used to properly place the link
-    yield { code, range };
-  }
-}
